requirements (7).txt
# Requirements - Emotion Recognition YOLOv8 + RAF-DB
# =====================================================
# Installer avec : pip install -r requirements.txt

# Deep Learning
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Application Web
streamlit>=1.28.0
plotly>=5.15.0

# Utilitaires
tqdm>=4.65.0
pyyaml>=6.0
